refactor: road to v3
