--- .eslintignore
@@ -1,5 +1,0 @@
-/coverage
-/dist
-/styleguide
-jest.config.js
-styleguide.config.js
