--- .eslintrc
@@ -1,6 +1,0 @@
-{
-    "extends": ["plugin:zillow/recommended", "plugin:zillow/jest"],
-    "env": {
-        "browser": true
-    }
-}
