+++ .gitignore
@@ -1,6 +1,3 @@
 /coverage
 /dist
-/styleguide
 node_modules
-.idea
-npm-debug.log*
