+++ .storybook/preview.ts
@@ -1,0 +1,10 @@
+import type { Preview } from '@storybook/react';
+import './global.css';
+
+const preview: Preview = {
+  parameters: {
+    actions: { argTypesRegex: '^on[A-Z].*' },
+  },
+};
+
+export default preview;
