+++ lint-staged.config.cjs
@@ -1,0 +1,4 @@
+module.exports = {
+  '*.{cjs,ts,tsx}': 'eslint --fix',
+  '*': 'prettier --ignore-unknown --write',
+};
