--- src/index.js
@@ -1,3 +1,0 @@
-import ReactSlider from './components/ReactSlider/ReactSlider';
-
-export default ReactSlider;
